# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/movement/filter": {
            "get": {
                "tags": [
                    "Movements"
                ],
                "summary": "Retrieve movements by month, year, and category",
                "operationId": "06ef4a959256d8142c985631856b672e",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Year to filter by",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2024
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Month to filter by",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category to filter by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Food"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of movements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "rowid": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "lc_movimento": {
                                                "type": "string",
                                                "example": "Movement details"
                                            },
                                            "ano": {
                                                "type": "integer",
                                                "example": 2024
                                            },
                                            "mes": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "categoria": {
                                                "type": "string",
                                                "example": "Food"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/movement": {
            "post": {
                "tags": [
                    "Movements"
                ],
                "summary": "Create a new movement",
                "operationId": "f898c67d6f1644a2370b2dd345e6d021",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dia": {
                                        "description": "Day of the movement",
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "mes": {
                                        "description": "Month of the movement",
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "ano": {
                                        "description": "Year of the movement",
                                        "type": "integer",
                                        "example": 2024
                                    },
                                    "tipo": {
                                        "description": "Type of the movement",
                                        "type": "string",
                                        "example": "Expense"
                                    },
                                    "categoria": {
                                        "description": "Category ID of the movement",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "descricao": {
                                        "description": "Description of the movement",
                                        "type": "string",
                                        "example": "Grocery shopping"
                                    },
                                    "valor": {
                                        "description": "Value of the movement",
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.75
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movement created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "dia": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "mes": {
                                            "type": "integer",
                                            "example": 8
                                        },
                                        "ano": {
                                            "type": "integer",
                                            "example": 2024
                                        },
                                        "tipo": {
                                            "type": "string",
                                            "example": "Expense"
                                        },
                                        "categoria": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "descricao": {
                                            "type": "string",
                                            "example": "Grocery shopping"
                                        },
                                        "valor": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 150.75
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Could not create movement"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}